requirements.txt
Faster_Whisper
LangSegment>=0.3.1
PyYAML
chardet
cn2an
fastapi
flash-attention
ffmpeg-python
funasr==1.0.0
g2p_en
gradio>=4.29
gradio_client
jieba
jieba_fast
librosa==0.9.2
modelscope==1.10.0
numba
numpy
onnxruntime
psutil
pydantic
pydub
pygame
pyopenjtalk
pyloudnorm
pypinyin
pytorch-lightning
requests
scipy
sentencepiece
soundfile
srt
tensorboard
tqdm
transformers
uvicorn
wordsegment
